+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - [client] Do not attempt to enable services at install
+- [client] Handle status subcommand when no arguments are passed
 - [dbuilder] Fix issue with the repo installer package build
 - [moriod] Detect non-interactive invovation in install script
 - [ui] Downloads page now lists the Repository Installer packages